feat: :sparkles: controladores y servicios de preguntas, respuestas y actividadAlumno
Implementados CRUD completos para ActividadAlumno, Pregunta y Respuesta, incluyendo validación de rol de maestro en Pregunta, Respuesta y General (aún no hay validaciones en ActividadAlumno).

Refs #35

+++ backend/src/main/java/com/cerebrus/actividad/GeneralServiceImpl.java
@@ -10,7 +10,9 @@ import com.cerebrus.TipoActGeneral;
 import com.cerebrus.pregunta.Pregunta;
 import com.cerebrus.tema.Tema;
 import com.cerebrus.tema.TemaRepository;
-
+import com.cerebrus.usuario.Maestro;
+import com.cerebrus.usuario.Usuario;
+import com.cerebrus.usuario.UsuarioService;
 import com.cerebrus.pregunta.PreguntaRepository;
 
 @Service
@@ -20,19 +22,26 @@ public class GeneralServiceImpl implements GeneralService {
     private final GeneralRepository generalRepository;
     private final TemaRepository temaRepository;
     private final PreguntaRepository preguntaRepository;
+    private final UsuarioService usuarioService;
 
     @Autowired
-    public GeneralServiceImpl(GeneralRepository generalRepository, TemaRepository temaRepository, PreguntaRepository preguntaRepository) {
+    public GeneralServiceImpl(GeneralRepository generalRepository, TemaRepository temaRepository, 
+        PreguntaRepository preguntaRepository, UsuarioService usuarioService) {
         this.generalRepository = generalRepository;
         this.temaRepository = temaRepository;
         this.preguntaRepository = preguntaRepository;
+        this.usuarioService = usuarioService;
     }
 
     @Override
     public General crearActGeneral(String titulo, String descripcion, Integer puntuacion, Long temaId, Boolean respVisible, String comentariosRespVisible) {
 
-        Tema tema = temaRepository.findById(temaId).orElseThrow(() -> new RuntimeException("El tema es incorrecto"));
+        Usuario u = usuarioService.findCurrentUser();
+        if (!(u instanceof Maestro)) {
+            throw new RuntimeException("Solo un maestro puede crear actividades");
+        }
 
+        Tema tema = temaRepository.findById(temaId).orElseThrow(() -> new RuntimeException("El tema de la actividad no existe"));
         General actividad = new General();
         actividad.setTitulo(titulo);
         actividad.setDescripcion(descripcion);
@@ -51,7 +60,14 @@ public class GeneralServiceImpl implements GeneralService {
     @Transactional
     public General crearTipoTest(String titulo, String descripcion, Integer puntuacion, Long temaId, Boolean respVisible, String comentariosRespVisible,
             List<Long> preguntasId) {
+
+        Usuario u = usuarioService.findCurrentUser();
+        if (!(u instanceof Maestro)) {
+            throw new RuntimeException("Solo un maestro puede crear actividades");
+        }
+
         General tipoTest = crearActGeneral(titulo, descripcion, puntuacion, temaId, respVisible, comentariosRespVisible);
+
         List<Pregunta> preguntas = preguntaRepository.findAllById(preguntasId);
         tipoTest.setPreguntas(preguntas);
         tipoTest.setTipo(TipoActGeneral.TEST);
@@ -61,11 +77,17 @@ public class GeneralServiceImpl implements GeneralService {
     @Override
     @Transactional
     public General readActividad(Long id){
-        return generalRepository.findByIdWithPreguntas(id).orElseThrow(() -> new RuntimeException("TipoTest no encontrado"));
+        return generalRepository.findByIdWithPreguntas(id).orElseThrow(() -> new RuntimeException("Actividad tipo test no encontrada"));
     }
 
     @Override
     public General updateActGeneral(Long id, String titulo, String descripcion, Integer puntuacion, Boolean respVisible, String comentariosRespVisible){
+        
+        Usuario u = usuarioService.findCurrentUser();
+        if (!(u instanceof Maestro)) {
+            throw new RuntimeException("Solo un maestro puede actualizar actividades");
+        }
+        
         General actividad = generalRepository.findById(id).orElseThrow(() -> new RuntimeException("Actividad no encontrada"));
         actividad.setTitulo(titulo);
         actividad.setDescripcion(descripcion);
@@ -88,6 +110,11 @@ public class GeneralServiceImpl implements GeneralService {
     @Transactional
     public General updateTipoTest(Long id, String titulo, String descripcion, Integer puntuacion, Boolean respVisible, String comentariosRespVisible, List<Long> preguntasId){
      
+        Usuario u = usuarioService.findCurrentUser();
+        if (!(u instanceof Maestro)) {
+            throw new RuntimeException("Solo un maestro puede actualizar actividades");
+        }
+
         General tipoTest = updateActGeneral(id, titulo, descripcion, puntuacion, respVisible, comentariosRespVisible);
         if(preguntasId != null){
             List<Pregunta> preguntas = preguntaRepository.findAllById(preguntasId);
@@ -101,6 +128,12 @@ public class GeneralServiceImpl implements GeneralService {
     @Override
     @Transactional
     public void deleteActividad(Long id) {  
+        
+        Usuario u = usuarioService.findCurrentUser();
+        if (!(u instanceof Maestro)) {
+            throw new RuntimeException("Solo un maestro puede eliminar actividades");
+        }
+        
         General general = generalRepository.findById(id).orElseThrow(() -> new RuntimeException("Actividad tipo test no encontrada"));
         generalRepository.delete(general);
     }
