+++ backend/src/main/java/com/cerebrus/respuesta/RespuestaServiceImpl.java
@@ -5,6 +5,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.cerebrus.pregunta.PreguntaRepository;
+import com.cerebrus.usuario.UsuarioService;
+import com.cerebrus.usuario.Usuario;
+import com.cerebrus.usuario.Maestro;
 import com.cerebrus.pregunta.Pregunta;
 
 @Service
@@ -13,18 +16,26 @@ public class RespuestaServiceImpl implements RespuestaService {
 
     private final RespuestaRepository respuestaRepository;
     private final PreguntaRepository preguntaRepository;
+    private final UsuarioService usuarioService;
 
     @Autowired
-    public RespuestaServiceImpl(RespuestaRepository respuestaRepository, PreguntaRepository preguntaRepository) {
+    public RespuestaServiceImpl(RespuestaRepository respuestaRepository, 
+        PreguntaRepository preguntaRepository, UsuarioService usuarioService) {
         this.respuestaRepository = respuestaRepository;
         this.preguntaRepository = preguntaRepository;
+        this.usuarioService = usuarioService;
     }
 
     @Override
     @Transactional
     public Respuesta crearRespuesta(String respuesta, String imagen, Boolean correcta, Long preguntaId) {
         
-        Pregunta pregunta = preguntaRepository.findById(preguntaId).orElseThrow(() -> new RuntimeException("La pregunta es incorrecta"));
+        Usuario u = usuarioService.findCurrentUser();
+        if (!(u instanceof Maestro)) {
+            throw new RuntimeException("Solo un maestro puede crear respuestas");
+        }
+
+        Pregunta pregunta = preguntaRepository.findById(preguntaId).orElseThrow(() -> new RuntimeException("La pregunta de la respuesta no existe"));
         
         Respuesta respuestaObj = new Respuesta();
         respuestaObj.setRespuesta(respuesta);
@@ -43,6 +54,12 @@ public class RespuestaServiceImpl implements RespuestaService {
     @Override
     @Transactional
     public Respuesta updateRespuesta(Long id, String respuesta, String imagen, Boolean correcta) {
+
+        Usuario u = usuarioService.findCurrentUser();
+        if (!(u instanceof Maestro)) {
+            throw new RuntimeException("Solo un maestro puede actualizar respuestas");
+        }
+
         Respuesta respuestaObj = respuestaRepository.findById(id).orElseThrow(() -> new RuntimeException("La respuesta no existe"));
         respuestaObj.setRespuesta(respuesta);
         respuestaObj.setImagen(imagen);
@@ -53,6 +70,12 @@ public class RespuestaServiceImpl implements RespuestaService {
     @Override
     @Transactional
     public void deleteRespuesta(Long id) {
+
+        Usuario u = usuarioService.findCurrentUser();
+        if (!(u instanceof Maestro)) {
+            throw new RuntimeException("Solo un maestro puede eliminar respuestas");
+        }
+
         Respuesta respuestaObj = respuestaRepository.findById(id).orElseThrow(() -> new RuntimeException("La respuesta no existe"));
         respuestaRepository.delete(respuestaObj);
     }
